refactor: Settings tabs use useFormState + shared SettingsSaveBar
- Move useFormState to @shared/hooks/ (was Pipelines-only); Pipelines re-exports for compat
- Create SettingsSaveBar shared component with useSaveStatus hook
- Refactor GeneralTab (486 → 454 lines): eliminate 8 individual change handlers,
  manual useState/useEffect sync, and duplicate save logic
- Refactor AgentTab (395 → 362 lines): eliminate 5 individual change handlers,
  manual useState/useEffect sync, and duplicate save logic
- Both tabs now use useFormState for state management and SettingsSaveBar for save UI
- Extract clamp() utility and DEFAULTS constants for clarity
- Zero functionality changes — all fields, validation, and behavior preserved

Resolves P1 items from React frontend technical debt audit.
